Added '-Force' to copy commands

+++ daily_driver.ps1
@@ -56,40 +56,42 @@ Start-Sleep -Seconds 1;
 
 $SambaProgramFiles = '\\GRIFFINUNRAID\programfiles'
 
-$IsLaptop = ($env:USERDOMAIN | Select-String 'LAPTOP')
-If ($IsLaptop) {
-	$RealTimeScript = 'BackupWindowsLaptop.ffs_real'
-	$BackupScript = 'BackupWindowsLaptop.ffs_batch'
-} else {
-	$RealTimeScript = 'BackupWindowsDesktop.ffs_real'
-	$BackupScript = 'BackupWindowsDesktop.ffs_batch'
-}
+If (Test-Path $SambaProgramFiles) {
+	$IsLaptop = ($env:USERDOMAIN | Select-String 'LAPTOP')
+	If ($IsLaptop) {
+		$RealTimeScript = 'BackupWindowsLaptop.ffs_real'
+		$BackupScript = 'BackupWindowsLaptop.ffs_batch'
+	} else {
+		$RealTimeScript = 'BackupWindowsDesktop.ffs_real'
+		$BackupScript = 'BackupWindowsDesktop.ffs_batch'
+	}
 
-$RealTimeScriptRemotePath = (Join-Path $SambaProgramFiles $RealTimeScript)
-$RealTimeScriptLocalPath = (Join-Path $env:ProgramData $RealTimeScript)
-$BackupScriptRemotePath = (Join-Path $SambaProgramFiles $BackupScript)
-$BackupScriptLocalPath = (Join-Path $env:ProgramData $BackupScript)
+	$RealTimeScriptRemotePath = (Join-Path $SambaProgramFiles $RealTimeScript)
+	$RealTimeScriptLocalPath = (Join-Path $env:ProgramData $RealTimeScript)
+	$BackupScriptRemotePath = (Join-Path $SambaProgramFiles $BackupScript)
+	$BackupScriptLocalPath = (Join-Path $env:ProgramData $BackupScript)
 
-If ((Test-Path $RealTimeScriptRemotePath) -And (Test-Path $BackupScriptRemotePath)) {
-	Copy-Item -Path $RealTimeScriptRemotePath -Destination $RealTimeScriptLocalPath
-	Copy-Item -Path $BackupScriptRemotePath -Destination $BackupScriptLocalPath
-}
+	If ((Test-Path $RealTimeScriptRemotePath) -And (Test-Path $BackupScriptRemotePath)) {
+		Copy-Item -Path $RealTimeScriptRemotePath -Destination $RealTimeScriptLocalPath -Force
+		Copy-Item -Path $BackupScriptRemotePath -Destination $BackupScriptLocalPath -Force
+	}
 
-$WallpaperRealTimeScript = 'MirrorCuratedSlideshowWallpaper.ffs_real'
-$WallpaperBatchScript = 'MirrorCuratedSlideshowWallpaper.ffs_batch'
+	$WallpaperRealTimeScript = 'MirrorCuratedSlideshowWallpaper.ffs_real'
+	$WallpaperBatchScript = 'MirrorCuratedSlideshowWallpaper.ffs_batch'
 
-$WallpaperRealTimeScriptRemotePath = (Join-Path $SambaProgramFiles $WallpaperRealTimeScript)
-$WallpaperRealTimeScriptLocalPath = (Join-Path $env:ProgramData $WallpaperRealTimeScript)
-$WallpaperBatchScriptRemotePath = (Join-Path $SambaProgramFiles $WallpaperBatchScript)
-$WallpaperBatchScriptLocalPath = (Join-Path $env:ProgramData $WallpaperBatchScript)
+	$WallpaperRealTimeScriptRemotePath = (Join-Path $SambaProgramFiles $WallpaperRealTimeScript)
+	$WallpaperRealTimeScriptLocalPath = (Join-Path $env:ProgramData $WallpaperRealTimeScript)
+	$WallpaperBatchScriptRemotePath = (Join-Path $SambaProgramFiles $WallpaperBatchScript)
+	$WallpaperBatchScriptLocalPath = (Join-Path $env:ProgramData $WallpaperBatchScript)
 
-If ((Test-Path $WallpaperRealTimeScriptRemotePath) -And (Test-Path $WallpaperBatchScriptRemotePath)) {
-	Copy-Item -Path $WallpaperRealTimeScriptRemotePath -Destination $WallpaperRealTimeScriptLocalPath
-	Copy-Item -Path $WallpaperBatchScriptRemotePath -Destination $WallpaperBatchScriptLocalPath
-}
+	If ((Test-Path $WallpaperRealTimeScriptRemotePath) -And (Test-Path $WallpaperBatchScriptRemotePath)) {
+		Copy-Item -Path $WallpaperRealTimeScriptRemotePath -Destination $WallpaperRealTimeScriptLocalPath -Force
+		Copy-Item -Path $WallpaperBatchScriptRemotePath -Destination $WallpaperBatchScriptLocalPath -Force
+	}
 
-If (Test-Path (Join-Path $SambaProgramFiles 'realtimesync.bat')) {
-	Copy-Item -Path (Join-Path $SambaProgramFiles 'realtimesync.bat') -Destination (Join-Path $env:ProgramData 'Microsoft\Windows\Start Menu\Programs\Startup\realtimesync.bat') -Force
+	If (Test-Path (Join-Path $SambaProgramFiles 'realtimesync.bat')) {
+		Copy-Item -Path (Join-Path $SambaProgramFiles 'realtimesync.bat') -Destination (Join-Path $env:ProgramData 'Microsoft\Windows\Start Menu\Programs\Startup\realtimesync.bat') -Force
+	}
 }
 
 Get-Content -Path $Boxstarter.Log | Select-String -Pattern '^Failures$' -Context 0, 2 >> (Join-Path $env:USERPROFILE 'Desktop\boxstarter-failures.log')
